requirements.txt
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone-client
fastapi
uvicorn
pandas
python-dotenv
